Simplify and fix UT2
I was making a slide for this example, and a) wanted to make it a bit
more concise, and b) realised I was really confused about it.

I was confused because this code doesn't actually produce an error - the
`matmul` is valid. Checking the source,
https://stackoverflow.com/questions/43067338/tensor-multiplication-in-tensorflow,
it looks like at _some_ point it didn't _used_ to be valid, but now the
semantics of `matmul` have changed so that it _is_ valid.

I think the real story here is "Wow, `matmul` does _not_ behave as you'd
expect it to". So I've tried modifying the code to reflect that.